converted server folder from .js to .ts
